requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
yfinance>=0.2.30
requests>=2.31.0
google-genai>=1.0.0
duckduckgo-search>=4.0.0
pyxirr>=0.10.0
numpy>=1.24.0
